+++ rate_limiter_system/server/store/store.go
@@ -3,5 +3,5 @@ package store
 import "context"
 
 type Store interface {
-	Eval(ctx context.Context, ipAddress string, capacity float64, refillRate int64) (*int64, error)
+	Eval(ctx context.Context, ipAddress string, capacity float64, refillRate int64) (*float64, error)
 }
